fix:SN改为rs232的波特率,支持scpi的RST复位功能,优化显示模块代码完成

+++ auxiliary/config_manager.cpp
@@ -56,7 +56,7 @@ bool ConfigManager::s_enableCANServer  = true;
 bool ConfigManager::s_enableUARTServer = true;
 bool ConfigManager::s_enableDisplay    = true;
 
-std::atomic<quint8> ConfigManager::s_remoteSt{0};
+std::atomic<int> ConfigManager::s_remoteSt{0};
 QSettings* ConfigManager::s_settings = nullptr;
 
 // global variable - Internal fixation
